% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_}
\alias{browse_}
\alias{browse_sysrev}
\alias{browse_article}
\title{open browser to sysrev project,article,etc.}
\usage{
browse_sysrev(pid)

browse_article(pid, aid)
}
\arguments{
\item{pid}{the project id to update}

\item{aid}{the article id within the project id to update}
}
\description{
open browser to sysrev project,article,etc.
}
