% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_article}
\alias{browse_article}
\title{view_article
opens the given project article in a browser}
\usage{
browse_article(pid, aid)
}
\arguments{
\item{pid}{project id to view}

\item{aid}{article id to view}
}
\value{
nothing
}
\description{
view_article
opens the given project article in a browser
}
