% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_sysrev}
\alias{browse_sysrev}
\title{browse_sysrev
open browser to sysrev project page}
\usage{
browse_sysrev(pid)
}
\arguments{
\item{pid}{syrev project id}
}
\description{
browse_sysrev
open browser to sysrev project page
}
