% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{clone_sysrev}
\alias{clone_sysrev}
\title{Create a new project based on an existing project}
\usage{
clone_sysrev(
  source.pid,
  target.name,
  add.articles = T,
  add.labels = T,
  add.members = F,
  add.answers = F,
  get_if_exists = T,
  token = get_srtoken()
)
}
\arguments{
\item{source.pid}{project id of the source project}

\item{target.name}{name of the new project}

\item{add.articles}{import the articles from source project?}

\item{add.labels}{import the labels from source project?}

\item{add.members}{import the members from source project?}

\item{add.answers}{import the answers from source project?}

\item{get_if_exists}{gets the sysrev instead of cloning if it already exists}

\item{token}{a token with write acces}
}
\description{
Create a new project based on an existing project
}
