% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{common_params}
\alias{common_params}
\title{common_params}
\arguments{
\item{pid}{the project id to update}

\item{aid}{the article id within the project id to update}

\item{lid}{the label id to update}

\item{token}{a token with write acces}
}
\description{
provides common argument descriptions
}
\keyword{internal}
