% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{copy_answers}
\alias{copy_answers}
\title{set the value of a given article+label}
\usage{
copy_answers(source.pid, target.pid)
}
\arguments{
\item{source.pid}{source project}

\item{target.pid}{target project}
}
\value{
list of modified articles
}
\description{
set the value of a given article+label
}
