% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_predictions}
\alias{create_predictions}
\title{create predictions for a sysrev project}
\usage{
create_predictions(pid, aid, lid, lbl.value, pred, token = get_srtoken())
}
\arguments{
\item{pid}{the project id to update}

\item{aid}{the article id within the project id to update}

\item{lid}{the label id to update}

\item{lbl.value}{a value to predict for the given aid and lid}

\item{pred}{a prediction (0-1) for the given aid,lid and lbl.value}

\item{token}{a token with write acces}
}
\value{
json describing result of call
}
\description{
create predictions for a sysrev project
}
