% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_source_entity}
\alias{create_source_entity}
\title{create_source_entity}
\usage{
create_source_entity(sid, title, entity, token = get_key())
}
\arguments{
\item{sid}{the datasource sid in which to place the entity}

\item{title}{a title for the entity}

\item{entity}{a list that can be converted to json via jsonlite}

\item{token}{sysrev private token. This is an access token for sysrev.com. You can find your personal token on your sysrev user page.}
}
\value{
eid the entity id
}
\description{
create_source_entity
}
