% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_source_insilica}
\alias{create_source_insilica}
\title{create_source_insilica}
\usage{
create_source_insilica(name, description, url, token = get_key())
}
\arguments{
\item{name}{name of the source}

\item{description}{description of the source}

\item{url}{a reference url for the source}

\item{token}{your private token}
}
\value{
the sid of the created source
}
\description{
create_source_insilica
}
