% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_source_pmids}
\alias{create_source_pmids}
\title{create a data source with pubmed ids (pmids)}
\usage{
create_source_pmids(pid, name, pmids, token = get_srtoken())
}
\arguments{
\item{pid}{the project id to update}

\item{name}{give your source a name}

\item{pmids}{the pubmed ids to import}

\item{token}{a token with write acces}
}
\value{
success message
}
\description{
create a data source with pubmed ids (pmids)
}
