% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.getArticleData.R
\name{datasource.getArticleData}
\alias{datasource.getArticleData}
\title{datasource.getArticleData
This will be deprecated.}
\usage{
datasource.getArticleData(
  article_ids,
  datasource,
  token = keyring::key_get("sysrev.token")
)
}
\arguments{
\item{article_ids}{datasource article ids to get}

\item{datasource}{the datasource to get}

\item{token}{a token with read priveleges}
}
\value{
gets a dataframe for arbitrary datasources from datasource service
}
\description{
datasource.getArticleData
This will be deprecated.
}
