% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.getPubmedArticles.R
\name{datasource.getPubmedArticles}
\alias{datasource.getPubmedArticles}
\title{datasource.getPubmedArticles}
\usage{
datasource.getPubmedArticles(
  datasource.ids,
  token = keyring::key_get("sysrev.token")
)
}
\arguments{
\item{datasource.ids}{article identifiers on datapub}

\item{token}{a sysrev token - defaults to keyring "sysrev.token"}
}
\value{
dataframe with datasource.id,abstract,authors,date,keywords,primary_title,secondary_title,update,year,url
}
\description{
datasource.getPubmedArticles
}
