% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{datasource.gql}
\alias{datasource.gql}
\title{datasource.gql
a simple wrapper for datasource.insilica.co}
\usage{
datasource.gql(
  query,
  token = get_srkey(),
  .url = "https://datasource.insilica.co/graphql"
)
}
\arguments{
\item{query}{the graphql query}

\item{token}{a sysrev token with read access to the given project}

\item{.url}{the url for datasource}
}
\value{
A dataframe
}
\description{
datasource.gql
a simple wrapper for datasource.insilica.co
}
