% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.risFileCitationsByIds.R
\name{datasource.risFileCitationsByIds}
\alias{datasource.risFileCitationsByIds}
\title{Title}
\usage{
datasource.risFileCitationsByIds(
  datasource.ids,
  token = keyring::key_get("sysrev.token")
)
}
\arguments{
\item{datasource.ids}{the ids to get ris information from}

\item{token}{a token with read priveleges}
}
\value{
dataframe for ris citations
}
\description{
Title
}
