% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R, R/tidy.R
\name{get_answers}
\alias{get_answers}
\alias{tidy_answers}
\title{get sysrev answers tibble with answers as rsr types}
\usage{
get_answers(pid, enabled_only = T, token = get_srtoken())

tidy_answers(
  answers,
  concordance = F,
  collapse = F,
  enabled_only = T,
  token = get_srtoken()
)
}
\arguments{
\item{pid}{a sysrev project id - sysrev.com/p/<pid>}

\item{enabled_only}{only get answers from enabled articles?}

\item{token}{a sysrev token with read access to the given project}

\item{answers}{answers from a `get_answers` call}

\item{concordance}{whether to compute concordance}

\item{collapse}{whether to remove user_ids and collapse answers}
}
\value{
a tbl with tidy sysrev answers
}
\description{
get sysrev answers tibble with answers as rsr types
}
