% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{get_answers_list}
\alias{get_answers_list}
\title{create a list of tidy answer tibbles
a special `basic` value in the list is populated by categorical boolean and string labels
every other value in the list represents a sysrev group label <rsr_group>}
\usage{
get_answers_list(pid)
}
\arguments{
\item{pid}{the project-id to export}
}
\value{
list of tibbles
}
\description{
create a list of tidy answer tibbles
a special `basic` value in the list is populated by categorical boolean and string labels
every other value in the list represents a sysrev group label <rsr_group>
}
