% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{get_answers_tidy}
\alias{get_answers_tidy}
\title{create a tidy version of sysrev answers}
\usage{
get_answers_tidy(pid, drop.resolved.discordant = F)
}
\arguments{
\item{pid}{the project-id to export}

\item{drop.resolved.discordant}{removes answers for articles that have a resolved alternative}
}
\value{
a tbl with tidy sysrev answers
}
\description{
create a tidy version of sysrev answers
}
