% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_article}
\alias{get_article}
\title{get_article}
\usage{
get_article(aid, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{aid}{the project to get articles from, i.e sysrev.com/p/<project_id>}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
get the basic article data from a sysrev project
}
