% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_articles}
\alias{get_articles}
\title{get_articles}
\usage{
get_articles(pid, token = get_srkey())
}
\arguments{
\item{pid}{the project to get articles from, i.e sysrev.com/p/<project_id>}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
get the basic article data from a sysrev project
}
