% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_entities}
\alias{get_entities}
\title{get_answers}
\usage{
get_entities(pid, token = get_srkey())
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\description{
get_answers
}
\concept{TODO this should really just be the same as get_articles?}
