% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_labels_tbl}
\alias{get_labels_tbl}
\title{get_labels_tbl
a simpler verison of get_labels}
\usage{
get_labels_tbl(pid, token = get_key())
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\description{
get_labels_tbl
a simpler verison of get_labels
}
