% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{get_or_create_sysrev}
\alias{get_or_create_sysrev}
\title{get a sysrev project by name, or create it.}
\usage{
get_or_create_sysrev(project_name, token = get_srkey())
}
\arguments{
\item{project_name}{the name of the project you want to create}

\item{token}{a sysrev token with read access to the given project}
}
\value{
a sysrev project id pid
}
\description{
get a sysrev project by name, or create it.
}
