% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_predictions}
\alias{get_predictions}
\title{get_predictions}
\usage{
get_predictions(pid, token = get_srkey())
}
\arguments{
\item{pid}{the project to get articles from, i.e sysrev.com/p/<project_id>}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
get the predictions for a project
}
