% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{get_project_concordance_options}
\alias{get_project_concordance_options}
\title{get_project_concordance_options}
\usage{
get_project_concordance_options(pid = NA)
}
\arguments{
\item{pid}{sysrev project_id, allows concordance_options to be looked up from sysrev.com}
}
\value{
a project_concordant_options list
}
\description{
get_project_concordance_options
}
\details{
concordance_options are used to compare two objects of the same type
}
