% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{get_sroptions}
\alias{get_sroptions}
\title{get_sroptions}
\usage{
get_sroptions(pid = NA, token = get_srtoken())
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\value{
a project_concordant_options list
}
\description{
get_sroptions
}
\details{
concordance_options are used to compare two objects of the same type
}
