% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{get_sysrev}
\alias{get_sysrev}
\title{get a sysrev with a name}
\usage{
get_sysrev(name, token = get_srtoken())
}
\arguments{
\item{name}{the sysrev to get metadata from}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
TODO need to move this back to get.R file and fix evaluation order
get a sysrev metadata
}
