% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_sysrevs}
\alias{get_sysrevs}
\title{get sysrevs owned by the given organization}
\usage{
get_sysrevs(oname, token = get_srtoken())
}
\arguments{
\item{oname}{organization name (cap sensitive for now)}

\item{token}{a token with write acces}
}
\description{
get sysrevs owned by the given organization
}
