% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_users}
\alias{get_users}
\title{get_users
get the users in a project}
\usage{
get_users(pid, token = get_srtoken())
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\description{
get_users
get the users in a project
}
