% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R, R/get_labels.R
\name{getAnnotations}
\alias{getAnnotations}
\title{Get User Annotations From Project}
\usage{
getAnnotations(project)

getAnnotations(project)
}
\arguments{
\item{project}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}
}
\description{
This function allows you to get user annotations from sysrev for a specific project.
The project must be public.

This function allows you to get user annotations from sysrev for a specific project.
The project must be public.
}
\examples{
getAnnotations(project=3144)
getAnnotations(project=3144)
}
\keyword{annotations}
