% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import_article_text}
\alias{import_article_text}
\title{import_article_text
creates a data source for the given sysrev project with articles generated from combined title and abstract}
\usage{
import_article_text(pid, title, abstract, token = get_srkey())
}
\arguments{
\item{pid}{the project for which to create a source}

\item{title}{titles for articles}

\item{abstract}{abstracts for articles}

\item{token}{a sysrev token with read access to the given project}
}
\value{
success message
}
\description{
import_article_text
creates a data source for the given sysrev project with articles generated from combined title and abstract
}
