% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import_pmids}
\alias{import_pmids}
\title{create a data source with pubmed ids (pmids)}
\usage{
import_pmids(pid, pmids, token = get_srtoken())
}
\arguments{
\item{pid}{the project for which to create a source}

\item{pmids}{the pubmed ids to import}

\item{token}{a sysrev token with read access to the given project}
}
\value{
success message
}
\description{
create a data source with pubmed ids (pmids)
}
