% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import_raw_articles}
\alias{import_raw_articles}
\title{import_raw_articles
a simple wrapper for rplumber.sysrev.com httr api requests}
\usage{
import_raw_articles(pid, articles, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{pid}{the service path see rplumber.sysrev.com/__docs__/#/}

\item{articles}{list(list(`primary-title`=<string>,`abstract`==<string>))}

\item{token}{a sysrev token with read access to the given project}
}
\value{
success message
}
\description{
import_raw_articles
a simple wrapper for rplumber.sysrev.com httr api requests
}
