% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken.R
\name{import_source_insilica}
\alias{import_source_insilica}
\title{import_source_insilica
imports a datasource.insilica.co dataset
TODO - fix this or abandon datasource?}
\usage{
import_source_insilica(pid, sid, token = get_srkey())
}
\arguments{
\item{pid}{the id of the target project}

\item{sid}{the id of the dataset to import}

\item{token}{sysrev private token. This is an access token for sysrev.com. You can find your personal token on your sysrev user page.}
}
\value{
success message
}
\description{
import_source_insilica
imports a datasource.insilica.co dataset
TODO - fix this or abandon datasource?
}
