% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-article-text.R
\name{importArticleText}
\alias{importArticleText}
\title{Get sysrev predictions for given project/label name/text}
\usage{
importArticleText(project, token, articles)
}
\arguments{
\item{project}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}
}
\description{
Get sysrev predictions for given project/label name/text
}
\examples{
getAnnotations(project=3144)
}
\keyword{annotations}
