% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_answers_unnested.R
\name{list_answers_unnested}
\alias{list_answers_unnested}
\title{retrieve clean table data containing Sysrev project answers}
\usage{
list_answers_unnested(
  pid,
  filter = NULL,
  filter_type = c("aid", "short_label"),
  token
)
}
\arguments{
\item{pid}{A sysrev project id - sysrev.com/p/<pid>}

\item{filter}{(Optional) aid or short_label from the given project to filter answers on}

\item{filter_type}{Type of value to filter on - either "aid" or "short_label"}

\item{token}{A sysrev token with read access to the given project}
}
\value{
A list containing a tibble for each answer
}
\description{
retrieve clean table data containing Sysrev project answers
}
\examples{
pid <- 117883
p_aid <- 14352862
token <- rstudioapi::askForSecret("Sysrev token with project access")

rsr_tables <- list_answers_unnested(pid, p_aid, "aid", token = token)
-----
pid <- 119108
label <- "Clinical_Chemistry"
token <- rstudioapi::askForSecret("Sysrev token with project access")

rsr_tables <- list_answers_unnested(pid, label, "short_label", token = token)
}
