% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_small.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{A small example version of the PBMC dataset from package SeuratObject}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\usage{
pbmc_small
}
\description{
A subsetted version of 10X Genomics' 3k PBMC dataset
}
\keyword{datasets}
