% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{project_concordance_options}
\alias{project_concordance_options}
\title{project_concordance_options}
\usage{
project_concordance_options(consensus.labels = list(), na.rm = T)
}
\arguments{
\item{consensus.labels}{which labels are required to have consensus?}

\item{na.rm}{should empty answers be discarded?}
}
\value{
a project_concordant_options list
}
\description{
project_concordance_options
}
\details{
concordance_options are used to compare two objects of the same type
}
