% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{set the value of a given article+label}
\usage{
review(
  pid,
  aid,
  lid,
  answer,
  change = T,
  resolve = F,
  raw.json = F,
  token = get_srtoken()
)
}
\arguments{
\item{pid}{the project id to update}

\item{aid}{the article id within the project id to update}

\item{lid}{the label id to update}

\item{answer}{the new value - this is an R object that can be parsed into json.}

\item{change}{treat this label update as a change (use default if you don't know what this means)}

\item{resolve}{treat this label update as resolve (use the default if you don't know what this means)}

\item{raw.json}{treat this as raw.json?}

\item{token}{a token with write acces}
}
\value{
true if successful
}
\description{
set the value of a given article+label
}
