% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplumber.R
\name{rplumber}
\alias{rplumber}
\title{sysrev.plumber.get
a simple wrapper for rplumber.sysrev.com httr api requests}
\usage{
rplumber(
  path,
  params = list(),
  action = httr::GET,
  token = keyring::key_get("sysrev.token")
)
}
\arguments{
\item{path}{the service path see rplumber.sysrev.com/__docs__/#/}

\item{params}{list of http parameters}

\item{action}{either httr::GET httr::POST or any other httr action}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
sysrev.plumber.get
a simple wrapper for rplumber.sysrev.com httr api requests
}
