% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsr-package.R
\docType{package}
\name{rsr-package}
\alias{rsr}
\alias{rsr-package}
\title{rsr: rsr}
\description{
an R client for sysrev.com. Comes packaged with some data analysis tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sysrev/rsr}
  \item \url{https://r.sysrev.com}
}

}
\author{
\strong{Maintainer}: Thomas Luechtefeld \email{tom@insilica.co}

}
\keyword{internal}
