% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsr.R
\name{rsr_departures}
\alias{rsr_departures}
\alias{rsr_submissions}
\alias{rsr_indicators}
\alias{rsr_demographics}
\title{Summary reports}
\usage{
rsr_departures(bureau, year, month)

rsr_submissions(bureau, year, month)

rsr_indicators(bureau, year, month)

rsr_demographics(bureau, year, month)
}
\arguments{
\item{bureau}{UNHCR bureau name as it appears in the RSR application (string)}

\item{year}{Year (numeric)}

\item{month}{Month (numeric)}
}
\value{
A tibble
}
\description{
Departure summary

Submission summary

Indicators summary

Demographics summary
}
