% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{service_list}
\alias{service_list}
\title{list services available for given project}
\usage{
service_list(pid)
}
\arguments{
\item{pid}{the project id to update}
}
\description{
list services available for given project
}
