% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{service_run}
\alias{service_run}
\title{Run a service within the context of a given project}
\usage{
service_run(pid, service, ...)
}
\arguments{
\item{pid}{the project id to update}

\item{service}{name of the service to run}
}
\description{
Run a service within the context of a given project
}
