% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{sroptions}
\alias{sroptions}
\title{sroptions}
\usage{
sroptions(consensus.labels = list(), na.rm = T)
}
\arguments{
\item{consensus.labels}{which labels are required to have consensus?}

\item{na.rm}{should empty answers be discarded?}
}
\value{
a project_concordant_options list
}
\description{
sroptions
}
\details{
concordance_options are used to compare two objects of the same type
}
\keyword{internal}
