% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{srplumber.post}
\alias{srplumber.post}
\title{sysrev.plumber.post
a simple wrapper for rplumber.sysrev.com httr api requests}
\usage{
srplumber.post(path, body, token, encode = "json")
}
\arguments{
\item{path}{the service path see rplumber.sysrev.com/__docs__/#/}

\item{token}{a sysrev token with read access to the given project}

\item{params}{list of http parameters}
}
\value{
A dataframe
}
\description{
sysrev.plumber.post
a simple wrapper for rplumber.sysrev.com httr api requests
}
\keyword{internal}
