% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{srtidy_answer}
\alias{srtidy_answer}
\title{tidy.answer
parses character vector json encodings of answers to rsr value types}
\usage{
srtidy_answer(answer, value_type, simplify = F)
}
\arguments{
\item{answer}{a json anwer to transform}

\item{value_type}{'boolean', 'categorical', 'string', or 'group'}

\item{simplify}{if true, unlists the result}
}
\description{
tidy.answer
parses character vector json encodings of answers to rsr value types
}
