% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysrev.getAnnotations.R
\name{sysrev.getAnnotations}
\alias{sysrev.getAnnotations}
\title{Get User Annotations From Project}
\usage{
sysrev.getAnnotations(project, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{project}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{your personal token, defaults to keyring::key_get("sysrev.token")}
}
\description{
This function allows you to get user annotations from sysrev for a specific project.
The project must be public.
}
\keyword{annotations}
