% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysrev.getArticles.R
\name{sysrev.getArticles}
\alias{sysrev.getArticles}
\title{sysrev.getArticles}
\usage{
sysrev.getArticles(project, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{project}{the project to get articles from, i.e sysrev.com/p/<project_id>}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
get the basic article data from a sysrev project
}
