% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysrev.getGroupLabelAnswers.R
\name{sysrev.getGroupLabelAnswers}
\alias{sysrev.getGroupLabelAnswers}
\title{Get user group label answers from project. returns a list with named dataframes corresponding to group label answers}
\usage{
sysrev.getGroupLabelAnswers(project_id, token = get_srkey())
}
\arguments{
\item{project_id}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\description{
Get user answers in a long form data frame
}
