% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysrev.getLabelAnswers.R
\name{sysrev.getLabelAnswers}
\alias{sysrev.getLabelAnswers}
\title{Get user answers from project}
\usage{
sysrev.getLabelAnswers(project, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{project}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{your personal token. get it with RSysrev::loginAPIToken()}
}
\description{
Get user answers in a long form data frame
}
