% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysrev.importDataset.R
\name{sysrev.importDataset}
\alias{sysrev.importDataset}
\title{sysrev.importDataset
Import a dataset from datasource as a source on SysRev. Result is true when import is successful,false otherwise.}
\usage{
sysrev.importDataset(
  datasource.id,
  project.id,
  token = keyring::key_get("sysrev.token")
)
}
\arguments{
\item{datasource.id}{the id of the dataset to import}

\item{project.id}{the sysrev project to import into}

\item{token}{a token with write priveleges}
}
\value{
true if the query works
}
\description{
sysrev.importDataset
Import a dataset from datasource as a source on SysRev. Result is true when import is successful,false otherwise.
}
