% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sysrev.webapi}
\alias{sysrev.webapi}
\title{sysrev.webapi
a simple wrapper for the sysrev.com/web-api/ service
documentation at https://sysrev.com/web-api/doc}
\usage{
sysrev.webapi(path, params = list(), token = get_key())
}
\arguments{
\item{path}{the service path see rplumber.sysrev.com/__docs__/#/}

\item{params}{list of http parameters}

\item{token}{a sysrev token with read access to the given project}
}
\value{
some json
}
\description{
sysrev.webapi
a simple wrapper for the sysrev.com/web-api/ service
documentation at https://sysrev.com/web-api/doc
}
