% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{test_token}
\alias{test_token}
\title{test_token}
\usage{
test_token(token)
}
\arguments{
\item{token}{sysrev private token. This is an access token for sysrev.com. You can find your personal token on your sysrev user page.}
}
\value{
the users sysrev token
}
\description{
test_token
}
