% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.answer}
\alias{tidy.answer}
\title{tidy.answer
parses character vector json encodings of answers to rsr value types}
\usage{
tidy.answer(value_type, answer)
}
\arguments{
\item{value_type}{'boolean', 'categorical', 'string', or 'group'}

\item{answer}{a json anwer to transform}
}
\description{
tidy.answer
parses character vector json encodings of answers to rsr value types
}
