% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.answers.group}
\alias{tidy.answers.group}
\title{tidy.answer}
\usage{
tidy.answers.group(answer)
}
\arguments{
\item{answer}{character vector json encoding for group labels}
}
\value{
<srr_group> vector which is a subclass of <tbl>
}
\description{
tidy.answer
}
\keyword{internal}
