% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{update_entity}
\alias{update_entity}
\title{update_entity
updates the primary title or abstract of an entity}
\usage{
update_entity(article_id, key, value, token = key_get("sysrev.token"))
}
\arguments{
\item{article_id}{the sysrev article_id to modify}

\item{key}{the key to modify (either 'primary-title' or 'abstract')}

\item{value}{what string value to update the key}

\item{token}{a sysrev token with read access to the given project}
}
\value{
a success message
}
\description{
update_entity
updates the primary title or abstract of an entity
}
