% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_articles.R
\name{view_articles}
\alias{view_articles}
\title{view_articles}
\usage{
view_articles(pid, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{pid}{the project to get articles from, i.e sysrev.com/p/<project_id>}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
get the basic article data from a sysrev project
}
