% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_project}
\alias{view_project}
\title{view_project
opens the given project in your browser}
\usage{
view_project(pid)
}
\arguments{
\item{pid}{project id to view}
}
\value{
nothing
}
\description{
view_project
opens the given project in your browser
}
