% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concordant}
\alias{concordant}
\title{concordant}
\usage{
concordant(a, single.concordance = T, pco)
}
\arguments{
\item{a}{input vector}

\item{single.concordance}{should length 1 value be considered concordant?}

\item{pco}{concordance settings from `get_sroptions(pid)`}
}
\value{
`concordant()` returns T when all answers are concordant
}
\description{
concordant
}
\details{
concordant refers to whether two sysrev answers are equivalent
}
\keyword{answer}
\keyword{comparison}
\keyword{sysrev}
