% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concordant2}
\alias{concordant2}
\title{check for binary answer concordance}
\usage{
concordant2(a, b, options)
}
\arguments{
\item{a}{first object to compare}

\item{b}{second object to compare}

\item{options}{additional arguments for specific answer subclasses}
}
\value{
`concordant()` returns T when a and b are 'concordant'
}
\description{
check for binary answer concordance
}
\details{
concordant refers to whether two sysrev answers are equivalent
TODO add a json equality to cover all undefined sysrev types
}
\keyword{answer}
\keyword{comparison}
\keyword{sysrev}
