% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_sysrev}
\alias{create_sysrev}
\title{create_sysrev
create a sysrev project}
\usage{
create_sysrev(name, get_if_exists = F, token = get_srtoken())
}
\arguments{
\item{name}{the name of the project you want to create}

\item{get_if_exists}{get the project if it already exists}

\item{token}{a sysrev token with read access to the given project}
}
\value{
json describing result of call
}
\description{
create_sysrev
create a sysrev project
}
