% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_}
\alias{get_}
\alias{get_articles}
\alias{get_predictions}
\alias{get_labels}
\alias{get_users}
\title{basic getters}
\usage{
get_articles(pid, token = get_srtoken())

get_predictions(pid, token = get_srtoken())

get_labels(pid, token = get_srtoken())

get_users(pid, token = get_srtoken())
}
\arguments{
\item{pid}{a sysrev project id - sysrev.com/p/<pid>}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A tibble
}
\description{
export tibbles from sysrev projects
}
