% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_answers}
\alias{get_answers}
\title{get_answers}
\usage{
get_answers(pid, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\description{
get_answers
}
