% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R, R/tidy.R
\name{get_answers}
\alias{get_answers}
\alias{get_answers_tidy}
\title{get sysrev answers tibble with answers as rsr types}
\usage{
get_answers(pid, token = get_srtoken())

get_answers_tidy(
  pid,
  concordance = F,
  concordant.collapse = F,
  token = get_srtoken()
)
}
\arguments{
\item{pid}{a sysrev project id - sysrev.com/p/<pid>}

\item{token}{a sysrev token with read access to the given project}

\item{conform}{remove user_ids and use a single answer for each aid+lid}
}
\value{
a tbl with tidy sysrev answers
}
\description{
get sysrev answers tibble with answers as rsr types
}
