% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R, R/tidy.R
\name{get_answers}
\alias{get_answers}
\alias{get_answers_tidy}
\alias{get_answers_list}
\title{get_answers}
\usage{
get_answers(pid, token = get_srtoken())

get_answers_tidy(pid, drop.resolved.discordant = F, token = get_srtoken())

get_answers_list(pid, token = get_srtoken())
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}

\item{drop.resolved.discordant}{remove answers for articles that have a resolved alternative?}
}
\value{
a tbl with tidy sysrev answers

list of tibbles
}
\description{
a special `basic` value in the list is populated by categorical boolean and string labels
every other value in the list represents a sysrev group label <rsr_group>
}
