% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{get_answers_list}
\alias{get_answers_list}
\title{get a list of tidy answer tibbles}
\usage{
get_answers_list(
  pid,
  concordance = F,
  collapse = F,
  token = get_srtoken(),
  ...
)
}
\arguments{
\item{pid}{a sysrev project id - sysrev.com/p/<pid>}

\item{concordance}{whether to compute concordance}

\item{collapse}{whether to remove user_ids and collapse answers}

\item{token}{a sysrev token with read access to the given project}
}
\value{
list of tibbles
}
\description{
a special `basic` value in the list is populated by categorical boolean and string labels
every other value in the list represents a sysrev group label <rsr_group>
}
