% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{get_answers_tidy}
\alias{get_answers_tidy}
\title{create a tidy version of sysrev answers}
\usage{
get_answers_tidy(pid, drop.resolved.discordant = F, token = get_srtoken())
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{drop.resolved.discordant}{remove answers for articles that have a resolved alternative?}

\item{token}{a sysrev token with read access to the given project}
}
\value{
a tbl with tidy sysrev answers
}
\description{
create a tidy version of sysrev answers
}
