% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_article}
\alias{get_article}
\title{get_article}
\usage{
get_article(aid, token = get_srtoken())
}
\arguments{
\item{aid}{the article to get data for}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
get the basic article data from a sysrev project
}
