% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_labels}
\alias{get_labels}
\title{get_labels
TODO this needs some refactoring
get the label definitions in a project}
\usage{
get_labels(pid, token = keyring::key_get("sysrev.token"))
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\description{
get_labels
TODO this needs some refactoring
get the label definitions in a project
}
