% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_labels}
\alias{get_labels}
\title{get label definitions}
\usage{
get_labels(pid, token = get_srtoken())
}
\arguments{
\item{pid}{The project identifier.  For sysrev.com/p/3144 the identifier is 3144}

\item{token}{a sysrev token with read access to the given project}
}
\description{
get label definitions
}
