% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{get_sroptions}
\alias{get_sroptions}
\title{get_sroptions}
\usage{
get_sroptions(pid = NA)
}
\arguments{
\item{pid}{sysrev project_id, allows concordance_options to be looked up from sysrev.com}
}
\value{
a project_concordant_options list
}
\description{
get_sroptions
}
\details{
concordance_options are used to compare two objects of the same type
}
