% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{get_sysrev}
\alias{get_sysrev}
\title{get_sysrev
TODO need to move this back to get.R file and fix evaluation order
get a sysrev metadata}
\usage{
get_sysrev(project_name, token = get_srkey())
}
\arguments{
\item{project_name}{the sysrev to get metadata from}

\item{token}{a sysrev token with read access to the given project}
}
\value{
A dataframe
}
\description{
get_sysrev
TODO need to move this back to get.R file and fix evaluation order
get a sysrev metadata
}
