% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R, R/create.R
\name{get_sroptions}
\alias{get_sroptions}
\alias{get_sysrev}
\title{get sysrev project/options with a name or id}
\usage{
get_sroptions(pid = NA, token = get_srtoken())

get_sysrev(name, pid = NA, token = get_srtoken())
}
\arguments{
\item{pid}{a sysrev project id - sysrev.com/p/<pid>}

\item{token}{a sysrev token with read access to the given project}

\item{name}{the sysrev to get metadata from}
}
\value{
A dataframe
}
\description{
get sysrev project/options with a name or id
}
