% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import_pmids}
\alias{import_pmids}
\title{import_pmids
creates a data source for the given sysrev project with the given pubme ids}
\usage{
import_pmids(pid, pmids, token = get_srkey())
}
\arguments{
\item{pid}{the project for which to create a source}

\item{pmids}{the pubmed ids to import}

\item{token}{a sysrev token with read access to the given project}
}
\value{
success message
}
\description{
import_pmids
creates a data source for the given sysrev project with the given pubme ids
}
