% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysrev.setLabelAnswer.R
\name{review}
\alias{review}
\title{review}
\usage{
review(
  project.id,
  article.id,
  lbl.id,
  lbl.value,
  token = keyring::key_get("sysrev.token")
)
}
\arguments{
\item{project.id}{the project id to update}

\item{article.id}{the article id within the project id to update}

\item{lbl.id}{the label id to update (see \link[RSysrev]{sysrev.getLabelDefinitions})}

\item{lbl.value}{the new value}

\item{token}{a token with write access}
}
\value{
true if successful
}
\description{
review
}
