% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{review}
\usage{
review(
  pid,
  aid,
  lid,
  lval,
  resolve = T,
  change = T,
  confirm = T,
  token = get_key()
)
}
\arguments{
\item{pid}{the project id to update}

\item{aid}{the article id within the project id to update}

\item{lid}{the label id to update (see \link[rsysrev]{get_labels})}

\item{lval}{the new value - this is an R object that can be parsed into json.}

\item{resolve}{treat this label update as resolve (use the default if you don't know what this means)}

\item{change}{treat this label update as a change (use default if you don't know what this means)}

\item{confirm}{treat this label update as a confirmed label (use default if you don't know what this means)}

\item{token}{a token with write acces}
}
\value{
true if successful
}
\description{
review
}
